server/requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
Deprecated==1.2.13
Flask==2.2.2
Flask-Cors==3.0.10
gcloud==0.18.3
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
oauth2client==4.1.3
protobuf==4.21.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pyparsing==3.0.9
Pyrebase4==4.6.0
python-dotenv==0.21.0
python-jwt==4.0.0
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
